Please commit your changes or stash them before you switch branches
