+++ index.html
@@ -227,6 +227,6 @@
       <button type="button" class="error__button">Попробовать снова</button>
     </div>
   </template>
-  <script defer type="text/javascript" src="js/main.js"></script>
+  <script defer type="module" src="js/main.js"></script>
 </body>
 </html>
